# Conflicts:
#	target/classes/Controller/Controller.class
#	target/classes/Controller/DataBase.class
#	target/classes/Models/Product.class
#	target/classes/Models/ProductStatus.class
#	target/classes/View/Menu/OffsAndProductsMenu/OffsPage$1.class
#	target/classes/View/Menu/OffsAndProductsMenu/OffsPage$2.class
#	target/classes/View/Menu/OffsAndProductsMenu/ProductsPage$1.class
#	target/classes/View/Menu/OffsAndProductsMenu/ProductsPage$2.class
#	target/classes/View/Menu/UserArea/CostumerArea/CostumerArea$2.class
#	target/classes/View/Menu/UserArea/CostumerArea/ViewCart$2.class
#	target/classes/View/Menu/UserArea/CostumerArea/ViewCart$3.class
#	target/classes/View/Menu/UserArea/CostumerArea/ViewCart$4.class
#	target/classes/View/Menu/UserArea/CostumerArea/ViewCart$5.class
#	target/classes/View/Menu/UserArea/ManagerArea/ViewDiscountCodes$1.class
#	target/classes/View/Menu/UserArea/ManagerArea/ViewDiscountCodes$2.class
#	target/classes/View/Menu/UserArea/ManagerArea/ViewDiscountCodes$3.class
#	target/classes/View/Menu/UserArea/ManagerArea/ViewDiscountCodes$4.class
#	target/classes/View/Menu/UserArea/SellerArea/SellerArea$1.class
#	target/classes/View/Menu/UserArea/SellerArea/SellerArea$2.class
#	target/classes/View/Menu/UserArea/SellerArea/SellerArea$3.class
#	target/classes/View/Menu/UserArea/SellerArea/SellerArea$4.class
#	target/classes/View/Menu/UserArea/SellerArea/SellerArea$5.class
#	target/classes/View/Menu/UserArea/SellerArea/SellerArea$6.class
#	target/classes/View/Menu/UserArea/SellerArea/SellerArea$7.class
#	target/classes/View/Menu/UserArea/SellerArea/ViewOffs$3.class
#	target/classes/View/Menu/UserArea/SellerArea/ViewOffs$4.class
#	target/test-classes/Controller/ControllerTest.class
#	target/test-classes/Controller/DataBaseTest.class
#	target/test-classes/Controller/OffAndProductMenuControllerTest.class

+++ src/main/java/View/Menu/UserArea/CostumerArea/CostumerArea.java
@@ -76,7 +76,7 @@ public class CostumerArea extends Menu {
                         View.printString("invalid amount");
                         continue;
                     }
-                    View.printString(CostumerAreaController.increaseCredit(Long.parseLong(credit)));
+                    View.printString(CostumerAreaController.increaseCredit(Double.parseDouble(credit)));
                 }
                 this.parentMenu.run(lastCommand);
             }
